requirements.txt
bottle==0.12.19
certifi==2020.11.8
chardet==3.0.4
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
idna==2.10
pyserial==3.5
requests==2.25.0
urllib3==1.26.2
